chore: MyMenuComponent 패키지 이동 (#31)
